+++ game/FootballSLG/Source/FootballSLG/MatchGameMode.cpp
@@ -1,0 +1,7 @@
+#include "MatchGameMode.h"
+#include "Footballer.h"
+
+AMatchGameMode::AMatchGameMode()
+{
+	DefaultPawnClass = AFootballer::StaticClass();
+}